requirements.txt
requests
websocket-client
rel
redis
